feat(core): Refactor auth with operator role, validation, and cluster/storage improvements
- Add RoleOperator and ErrLastAdmin sentinel error
- Add auth config nil check and session validation
- Add username/role validation and duplicate user detection
- Improve cluster/raft hardstate and snapshot handling
- Improve storage WAL with rotation and compaction
- Improve server handler error handling and timeouts
- Add coverage test helpers
- Improve DoH/DoQ/ODoH handlers

+++ internal/cache/cache_test.go
@@ -476,7 +476,7 @@ func TestEntryRemainingTTL(t *testing.T) {
 	entry.ExpireTime = now.Add(time.Duration(int64(^uint32(0))+1) * time.Second)
 	remaining = entry.RemainingTTL(now)
 	if remaining != ^uint32(0) {
-		t.Errorf("expected saturated remaining TTL %d, got %d", uint32(^uint32(0)), remaining)
+		t.Errorf("expected saturated remaining TTL %d, got %d", ^uint32(0), remaining)
 	}
 }
